requirements.txt
streamlit==0.82.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.3.1
